Python_scripts/Get_challengers_euw/requirements.txt
pandas
requests
google-cloud-storage
google-auth